+++ src/Functions.Worker.Extensions.OpenAI/Embeddings/EmbeddingsOptionsJsonConverter.cs
@@ -4,25 +4,25 @@
 using System.ClientModel.Primitives;
 using System.Text.Json;
 using System.Text.Json.Serialization;
-using Azure.AI.OpenAI;
+using OpenAI.Embeddings;
 
 namespace Microsoft.Azure.Functions.Worker.Extensions.OpenAI.Embeddings;
 
 /// <summary>
 /// EmbeddingsOptions JSON converter needed to serialize and deserialize the EmbeddingsOptions object with the dotnet worker.
 /// </summary>
-class EmbeddingsOptionsJsonConverter : JsonConverter<EmbeddingsOptions>
+class EmbeddingsOptionsJsonConverter : JsonConverter<EmbeddingGenerationOptions>
 {
     static readonly ModelReaderWriterOptions JsonOptions = new("J");
 
-    public override EmbeddingsOptions Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+    public override EmbeddingGenerationOptions Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
     {
         using JsonDocument jsonDocument = JsonDocument.ParseValue(ref reader);
-        return ModelReaderWriter.Read<EmbeddingsOptions>(BinaryData.FromString(jsonDocument.RootElement.GetRawText()))!;
+        return ModelReaderWriter.Read<EmbeddingGenerationOptions>(BinaryData.FromString(jsonDocument.RootElement.GetRawText()))!;
     }
 
-    public override void Write(Utf8JsonWriter writer, EmbeddingsOptions value, JsonSerializerOptions options)
+    public override void Write(Utf8JsonWriter writer, EmbeddingGenerationOptions value, JsonSerializerOptions options)
     {
-        ((IJsonModel<EmbeddingsOptions>)value).Write(writer, JsonOptions);
+        ((IJsonModel<EmbeddingGenerationOptions>)value).Write(writer, JsonOptions);
     }
 }