+++ java-library/src/main/java/com/microsoft/azure/functions/openai/annotation/assistant/AssistantMessage.java
@@ -12,31 +12,20 @@ package com.microsoft.azure.functions.openai.annotation.assistant;
  * the chat agent, and the name of the calling function if applicable.
  * </p>
  */
-public class ChatMessage {
+public class AssistantMessage {
 
     private String content;
     private String role;
     private String toolCalls;
 
     /**
-     * Initializes a new instance of the ChatMessage class.
-     *
-     * @param content The content of the message.
-     * @param role    The role of the chat agent.
-     */
-    public ChatMessage(String content, String role) {
-        this.content = content;
-        this.role = role;
-    }
-
-    /**
-     * Initializes a new instance of the ChatMessage class.
+     * Initializes a new instance of the AssistantMessage class.
      *
      * @param content   The content of the message.
      * @param role      The role of the chat agent.
      * @param toolCalls The toolCalls of the calling function if applicable.
      */
-    public ChatMessage(String content, String role, String toolCalls) {
+    public AssistantMessage(String content, String role, String toolCalls) {
         this.content = content;
         this.role = role;
         this.toolCalls = toolCalls;
@@ -90,7 +79,7 @@ public class ChatMessage {
     /**
      * Sets the toolCalls of the calling function if applicable.
      * 
-     * @param name The toolCalls of the calling function if applicable.
+     * @param toolCalls The toolCalls of the calling function if applicable.
      */
     public void setToolCalls(String toolCalls) {
         this.toolCalls = toolCalls;