+++ CHANGELOG.md
@@ -11,8 +11,14 @@ Starting v0.1.0 for Microsoft.Azure.WebJobs.Extensions.OpenAI.AzureAISearch, it 
 
 ## v0.19.0 - Unreleased
 
+### Breaking
+
+- Model properties named to ChatModel and EmbeddingsModel in related bindings
+- Managed identity support through config section and binding parameter AIConnectionName.
+
 ### Changed
 
+- Updated Azure.AI.OpenAI from 1.0.0-beta.15 to 2.1.0
 - Updated Azure.Data.Tables from 12.9.1 to 12.10.0, Azure.Identity from 1.12.1 to 1.13.2, Microsoft.Extensions.Azure from 1.7.5 to 1.8.0
 
 ## v0.18.0 - 2024/10/08